requirements.txt
# Telegram
telethon==1.34.0
python-telegram-bot==20.7

# Discord
discord.py==2.3.2

# Utility
python-dotenv==1.0.0
asyncio==3.4.3
colorlog==6.7.0

# Optional (untuk formatting lebih baik)
rich==13.7.0

# added for WIN
pynacl==1.6.2
cryptg==0.5.2
